refactor: 1차 종합 리팩토링
