+++ src/main/java/com/cinefinder/brand/data/entity/Brand.java
@@ -1,4 +1,4 @@
-package com.cinefinder.favorite.data;
+package com.cinefinder.brand.data.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -11,15 +11,13 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@Table(name = "FAVORITE")
-public class Favorite {
+@Table(name = "BRAND", uniqueConstraints = {@UniqueConstraint(columnNames = "name")})
+public class Brand {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false)
-    private Long userId;
-
-    @Column(nullable = false)
-    private Long movieId;
+    @Column(nullable = false, unique = true)
+    private String name;
 }
