+++ src/main/java/com/cinefinder/brand/data/repository/BrandRepository.java
@@ -1,6 +1,6 @@
 package com.cinefinder.brand.data.repository;
 
-import com.cinefinder.brand.data.Brand;
+import com.cinefinder.brand.data.entity.Brand;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
