+++ src/main/java/com/cinefinder/movie/controller/MovieController.java
@@ -3,7 +3,6 @@ package com.cinefinder.movie.controller;
 import com.cinefinder.global.mapper.ResponseMapper;
 import com.cinefinder.global.response.BaseResponse;
 import com.cinefinder.global.util.statuscode.ApiStatus;
-import com.cinefinder.movie.data.Movie;
 import com.cinefinder.movie.data.model.BoxOffice;
 import com.cinefinder.movie.data.model.MovieDetails;
 import com.cinefinder.movie.service.MovieService;
