+++ src/main/java/com/cinefinder/movie/data/dto/SimplifiedMovieDto.java
@@ -1,4 +1,4 @@
-package com.cinefinder.movie.data;
+package com.cinefinder.movie.data.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
