+++ src/main/java/com/cinefinder/movie/data/repository/MovieRepository.java
@@ -1,6 +1,6 @@
 package com.cinefinder.movie.data.repository;
 
-import com.cinefinder.movie.data.Movie;
+import com.cinefinder.movie.data.entity.Movie;
 import io.lettuce.core.dynamic.annotation.Param;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -11,6 +11,7 @@ import java.util.Optional;
 
 @Repository
 public interface MovieRepository extends JpaRepository<Movie, Long> {
+    // TODO : 업보
     Optional<Movie> findByMovieKey(String movieKey);
     Movie findByCgvCode(String cgvCode);
     Movie findByLotteCinemaCode(String lotteCode);
