+++ src/main/java/com/cinefinder/theater/data/entity/ElasticsearchTheater.java
@@ -1,4 +1,4 @@
-package com.cinefinder.theater.data;
+package com.cinefinder.theater.data.entity;
 
 import org.springframework.data.annotation.Id;
 import org.springframework.data.elasticsearch.annotations.Document;
