+++ src/main/java/com/cinefinder/theater/service/MegaboxTheaterCrawlerServiceImpl.java
@@ -2,7 +2,7 @@ package com.cinefinder.theater.service;
 
 import com.cinefinder.global.exception.custom.CustomException;
 import com.cinefinder.global.util.statuscode.ApiStatus;
-import com.cinefinder.theater.data.Theater;
+import com.cinefinder.theater.data.entity.Theater;
 import com.cinefinder.brand.data.repository.BrandRepository;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
