+++ src/main/java/com/cinefinder/theater/service/TheaterSyncService.java
@@ -1,6 +1,6 @@
 package com.cinefinder.theater.service;
 
-import com.cinefinder.theater.data.Theater;
+import com.cinefinder.theater.data.entity.Theater;
 import com.cinefinder.theater.data.repository.ElasticsearchTheaterRepository;
 import com.cinefinder.theater.data.repository.TheaterRepository;
 import lombok.RequiredArgsConstructor;
